+++ packages/network/test/browser/BrowserWebRtcConnection.test.ts
@@ -12,7 +12,9 @@ const connectionOpts1: ConstructorOptions = {
     iceServers: [],
     pingInterval: 5000,
     messageQueue: new MessageQueue<string>(TEST_CONFIG.webrtcSendBufferMaxMessageCount),
-    deferredConnectionAttempt: new DeferredConnectionAttempt()
+    deferredConnectionAttempt: new DeferredConnectionAttempt(),
+    portRange: TEST_CONFIG.webrtcPortRange,
+    maxMessageSize: TEST_CONFIG.webrtcMaxMessageSize
 }
 
 const connectionOpts2: ConstructorOptions = {
@@ -22,7 +24,9 @@ const connectionOpts2: ConstructorOptions = {
     iceServers: [],
     pingInterval: 5000,
     messageQueue: new MessageQueue<string>(TEST_CONFIG.webrtcSendBufferMaxMessageCount),
-    deferredConnectionAttempt: new DeferredConnectionAttempt()
+    deferredConnectionAttempt: new DeferredConnectionAttempt(),
+    portRange: TEST_CONFIG.webrtcPortRange,
+    maxMessageSize: TEST_CONFIG.webrtcMaxMessageSize
 }
 
 describe('BrowserWebRtcConnection', () => {